push man

+++ case_study_module_4/src/main/java/org/example/case_study_module_4/repository/MediaRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface MediaRepository extends CrudRepository<Media, Integer> {
+public interface MediaRepository extends CrudRepository<Media, Long> {
     @Query("SELECT m FROM Media m WHERE m.post.id = :postId")
     List<Media> findByPostId(@Param("postId") Long postId);
 }
